Implement freeze subcommand for arteacle class! v0.1.2
Implement a freeze subcommand that freezes the current state of the
arteacle class into a single file with a specififed git commit ID.

This will be particularly useful for knowing at which version I finish
a project at.

In future, it might be nice to make this subcommand optionally accept
a commit hash, and to create a frozen class at that commit.