Fix bug in resource name while fetching remote resource!  v1.8.1
In version 1.8.0 we refactored according to a recent version of
tex-macros (see jakewilliami/tex-macros#51).  As such, we need to
sometimes strip superfluous information from the file path.  We keep
the user and repo name in the path for information (e.g., in case we
change the repository name or even use a different user's macros).  I
updated this in freeze but not in fetch_resource_remote.