Implement letter option!  v1.9.0
This version introduces a letter option, using the lettear class
defined in jakewilliami/tex-macros.  This class was pre-existing but
is not used by me often (I mostly do hand-written letters) so this was
not a priority.  Closes #18.

In addition to this, we deprecate the --class option in favour of
--article, show help when no arguments are provided (#19), error for
invalid command line options, gracefuly handle when the target file
already exists (#14), and improve logging when checking sync status of
sourced files within each class (which is a process we should log
since it was introduced in v1.8.0).