Add latest tex-macros version to template!  v1.9.3
When we use mktex, we copy a base template to the target
directory (and also make sure required resources to use that are
updated locally).  This version modifies the base template before
writing it to the target directory, adding the current version of the
LaTeX configs (i.e., jakewilliami/tex-macros) that you are using.

We also clean up a bit with clippy.